HasDefaultTransactions: no more
